Add files via upload
This repository is for my first task as an intern in "BHARAT INTERN". The task is to create a machine learning model to predict house price using linear regression only. (Here I created both Linear regression and random forest Model for better understanding of accuracy) The code is done in Jupyter notebook.